Remove INCREMENT opcode
INCREMENT has turned out to be kind of an awkward opcode.

First, there were some questions about how to translate it from Nock,
where atoms are natural numbers, to this, where atoms are byte strings.
For example, on overflow, should the bytestring wrap back to 0, or grow?
Should leading 0s get trimmed? If it grows on overflow, is it better to
always grow to to make a sort of timing attack by having the behavior
depend on the contents of the atom?

Second, without jets, INCREMENT cannot really be said to be a useful
building block for ooperations like decrementing, adding, multiplying,
etc. It is really only good for small counters, and for that, iterating
the hash function works. Appending to a list would also work. It was
straightforward to rewrite the guessing_game testcase in terms of iterated
hash functions.