Switch endian-ness of numbers
This makes is consistent with byte-literals in the normal way of writing them.
In the literal x1234, the most significant byte is x12, now. This is also nice
because the nibbles are descending order now when written out. This endian-ness
is unusual in computing, but having the consistency between bytes written out and
math outcomes seems worth it.